refactor(ui): rename RouteLabel to RouteCountBadge and move under badge/
The component grew beyond a "label" primitive: it now draws a framed
capsule with inverted count colors, wraps a LabelCountBadge, and fits
squarely in the badge family alongside RouteBadge / AgencyBadge. Move
it to src/components/badge/ and rename to RouteCountBadge so the
name reflects its purpose (domain adapter yielding a label + count
badge for a route) and its location reflects its role.

- Rename src/components/label/route-label.tsx -> src/components/badge/route-count-badge.tsx
- Rename src/components/label/route-label.stories.tsx -> src/components/badge/route-count-badge.stories.tsx
- Rename component / interface / storybook title accordingly
- Update TimetableMetadata import and usage
- Update BaseLabel.stories comment reference
- Update CHANGELOG Unreleased entry
