+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [CalVer](https://calver.org/).
 
 ### Added
 
-- `RouteLabel` (新規、`src/components/label/route-label.tsx`): `BaseLabel` を wrap した display-only のルート表示ラベル。`TimetableMetadata` で `PillButton` の route 内訳行と並べて視認比較するために追加。本命の `BaseBadge` 実装前の試験導入。
+- `RouteCountBadge` (新規、`src/components/badge/route-count-badge.tsx`): route の display name と `route_color` を `LabelCountBadge` に橋渡しする domain adapter。`TimetableMetadata` の route 内訳表示を PillButton から置き換え。
 - `BaseLabel` に `style` prop を追加。GTFS の `route_color` のようなランタイム hex 値を inline style で渡せるように (既存 `PillButton` / `RouteBadge` と同じパターン)。
 
 ### Changed
